Add Node.js tests, snippets, and content
