+++ examples/node/v12/config.ts
@@ -1,4 +1,5 @@
 const APP_ID = "js-flexible-oseso";
+const EDGE_APP_ID = "sync-edge-server-cskhoow";
 const PBS_APP_ID = "example-testers-kvjdy";
 
-export { APP_ID, PBS_APP_ID };
+export { APP_ID, EDGE_APP_ID, PBS_APP_ID };
