+++ examples/node/v12/package.json
@@ -15,7 +15,7 @@
   "license": "ISC",
   "dependencies": {
     "fs-extra": "^11.1.1",
-    "realm": "^12.3.0"
+    "realm": "^12.8.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
