+++ source/sdk/node/app-services/connect-to-app-services-backend.txt
@@ -136,3 +136,99 @@ To encrypt App metadata:
       :tabid: javascript
 
       .. include:: /examples/generated/node/v12/formatted/metadata.test.snippet.encrypt-metadata.js.rst
+
+Connect to a Specific Server
+----------------------------
+
+By default, Atlas Device SDK connects to Atlas using the global ``baseURL``
+of ``https://services.cloud.mongodb.com``. In some cases, you may want to
+connect to a different server:
+
+- Your App Services App uses :ref:`local deployment <local-deployment>`, and 
+  you want to connect directly to a local ``baseURL`` in your region
+- You want to connect to an :ref:`Edge Server instance <edge-server-connect-from-client>`
+
+You can specify a ``baseURL`` in the 
+:js-sdk:`AppConfiguration <types/Realm.AppConfiguration.html>`:
+
+.. tabs-realm-languages::
+
+   .. tab::
+      :tabid: typescript
+
+      .. literalinclude:: /examples/generated/node/v12/formatted/app-services.test.snippet.custom-base-url.ts.rst
+        :language: typescript
+
+   .. tab::
+      :tabid: javascript
+
+      .. literalinclude:: /examples/generated/node/v12/formatted/app-services.test.snippet.custom-base-url.js.rst
+        :language: javascript
+
+Connect to a Different Server During Runtime
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. versionadded:: 12.8.0
+
+In some cases, you might want to change the ``baseURL`` while the app is
+running. For example, you might want to roam between Edge Servers, or
+move from an App Services connection to an Edge Server connection. 
+
+To change the ``baseURL`` during runtime, call the ``app.updateBaseUrl()``
+method. Note that the string value can't end in a trailing slash.
+
+.. tabs-realm-languages::
+
+   .. tab::
+      :tabid: typescript
+
+      .. literalinclude:: /examples/generated/node/v12/formatted/app-services.test.snippet.change-base-url.ts.rst
+        :language: typescript
+
+   .. tab::
+      :tabid: javascript
+
+      .. literalinclude:: /examples/generated/node/v12/formatted/app-services.test.snippet.change-base-url.js.rst
+        :language: javascript
+
+This API is experimental, so you must use the experimental import in the file
+where you want to use this API:
+
+.. literalinclude:: /examples/generated/node/v12/formatted/app-services.test.snippet.import-experimental-baseurl.ts.rst
+  :language: typescript
+
+TypeScript Config for Experimental API
+``````````````````````````````````````
+
+The ``app.updateBaseUrl()`` method is in an experimental module that you must
+import separately. To get the proper types from this module, you need to set the
+following fields in your ``tsconfig.json`` file:
+
+.. code-block:: json
+   :copyable: true
+
+   {
+      "target": "es2022",
+      "module": "node16",
+      "moduleResolution": "node16",
+      // ...
+   }
+
+
+Change Server with Logged-in user
+`````````````````````````````````
+
+If you want to change the ``baseURL`` after you have logged in a user and
+have opened a synced database, the app must perform a 
+:ref:`client reset <node-perform-a-client-reset>`. Perform these steps in your
+code:
+
+1. :ref:`Pause the Sync session <node-pause-or-resume-a-sync-session>`
+2. Update the ``baseURL`` by calling the ``app.updateBaseUrl()`` method
+3. Go through the authentication flow again to log in the user through the
+   new ``baseURL``
+4. Open a synced database pulling data from the new server
+
+Both the server and the client must be online for the user to authenticate and
+connect to the new server. If the server is not online or the client does not
+have a network connection, the user cannot authenticate and open the database.
